requirements.txt
fastapi
uvicorn
celery
redis
python-multipart
Pillow
numpy
streamlit
requests
gigachat
g4f
